+++ smells/main.go
@@ -5,6 +5,7 @@ import (
 
 	"openterps/dbconnector"
 	"openterps/models"
+	"openterps/simpleauth"
 
 	"github.com/gin-gonic/gin"
 )
@@ -25,53 +26,65 @@ type smellsInput struct {
 // POST /smells
 // Create a smells
 func CreateSmells(c *gin.Context) {
-	// Validate input
-	var input smellsInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
 
-	// Create smells
-	smells := models.Smell{
-		Name: input.Name,
-	}
-	dbconnector.DB.Create(&smells)
+	if userIsLoggedIn {
+		// Validate input
+		var input smellsInput
+		if err := c.ShouldBindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
 
-	c.JSON(http.StatusOK, gin.H{"data": smells})
+		// Create smells
+		smells := models.Smell{
+			Name: input.Name,
+		}
+		dbconnector.DB.Create(&smells)
+
+		c.JSON(http.StatusOK, gin.H{"data": smells})
+	}
 }
 
 // PATCH /smells/:id
 // Update a smells
 func UpdateSmells(c *gin.Context) {
-	var smells models.Smell
-	if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&smells).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Smell not found!"})
-		return
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
+
+	if userIsLoggedIn {
+		var smells models.Smell
+		if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&smells).Error; err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Smell not found!"})
+			return
+		}
+
+		// Validate input
+		var input smellsInput
+		if err := c.ShouldBindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
+		dbconnector.DB.Model(&smells).Updates(input)
+
+		c.JSON(http.StatusOK, gin.H{"data": smells})
 	}
-
-	// Validate input
-	var input smellsInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	dbconnector.DB.Model(&smells).Updates(input)
-
-	c.JSON(http.StatusOK, gin.H{"data": smells})
 }
 
 // DELETE /smells/:id
 // Delete a smells
 func DeleteSmells(c *gin.Context) {
-	var smells models.Smell
-	if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&smells).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Smells not found!"})
-		return
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
+
+	if userIsLoggedIn {
+		var smells models.Smell
+		if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&smells).Error; err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Smells not found!"})
+			return
+		}
+
+		dbconnector.DB.Delete(&smells)
+
+		c.JSON(http.StatusOK, gin.H{"data": true})
 	}
-
-	dbconnector.DB.Delete(&smells)
-
-	c.JSON(http.StatusOK, gin.H{"data": true})
 }
