+++ tastes/main.go
@@ -5,6 +5,7 @@ import (
 
 	"openterps/dbconnector"
 	"openterps/models"
+	"openterps/simpleauth"
 
 	"github.com/gin-gonic/gin"
 )
@@ -25,53 +26,65 @@ type tastesInput struct {
 // POST /tastes
 // Create a tastes
 func CreateTastes(c *gin.Context) {
-	// Validate input
-	var input tastesInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
 
-	// Create tastes
-	tastes := models.Taste{
-		Name: input.Name,
-	}
-	dbconnector.DB.Create(&tastes)
+	if userIsLoggedIn {
+		// Validate input
+		var input tastesInput
+		if err := c.ShouldBindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
 
-	c.JSON(http.StatusOK, gin.H{"data": tastes})
+		// Create tastes
+		tastes := models.Taste{
+			Name: input.Name,
+		}
+		dbconnector.DB.Create(&tastes)
+
+		c.JSON(http.StatusOK, gin.H{"data": tastes})
+	}
 }
 
 // PATCH /tastes/:id
 // Update a tastes
 func UpdateTastes(c *gin.Context) {
-	var tastes models.Taste
-	if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&tastes).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Taste not found!"})
-		return
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
+
+	if userIsLoggedIn {
+		var tastes models.Taste
+		if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&tastes).Error; err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Taste not found!"})
+			return
+		}
+
+		// Validate input
+		var input tastesInput
+		if err := c.ShouldBindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
+		dbconnector.DB.Model(&tastes).Updates(input)
+
+		c.JSON(http.StatusOK, gin.H{"data": tastes})
 	}
-
-	// Validate input
-	var input tastesInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	dbconnector.DB.Model(&tastes).Updates(input)
-
-	c.JSON(http.StatusOK, gin.H{"data": tastes})
 }
 
 // DELETE /tastes/:id
 // Delete a tastes
 func DeleteTastes(c *gin.Context) {
-	var tastes models.Taste
-	if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&tastes).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Tastes not found!"})
-		return
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
+
+	if userIsLoggedIn {
+		var tastes models.Taste
+		if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&tastes).Error; err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Tastes not found!"})
+			return
+		}
+
+		dbconnector.DB.Delete(&tastes)
+
+		c.JSON(http.StatusOK, gin.H{"data": true})
 	}
-
-	dbconnector.DB.Delete(&tastes)
-
-	c.JSON(http.StatusOK, gin.H{"data": true})
 }
