+++ terpenes/main.go
@@ -6,6 +6,7 @@ import (
 
 	"openterps/dbconnector"
 	"openterps/models"
+	"openterps/simpleauth"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -43,57 +44,69 @@ func CreateTerpene(c *gin.Context) {
 		return
 	}
 
-	// Create terpene
-	terpene := models.Terpene{
-		Name:       input.Name,
-		Categories: input.Category,
-		Tastes:     input.Taste,
-		Smells:     input.Smell,
-		Strains:    input.Strain,
-		Effects:    input.Effect,
-	}
-	dbconnector.DB.Create(&terpene)
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
 
-	c.JSON(http.StatusOK, gin.H{"data": terpene})
+	if userIsLoggedIn {
+		// Create terpene
+		terpene := models.Terpene{
+			Name:       input.Name,
+			Categories: input.Category,
+			Tastes:     input.Taste,
+			Smells:     input.Smell,
+			Strains:    input.Strain,
+			Effects:    input.Effect,
+		}
+		dbconnector.DB.Create(&terpene)
+
+		c.JSON(http.StatusOK, gin.H{"data": terpene})
+	}
 }
 
 // PATCH /terpenes/:id
 // Update a terpene
 func UpdateTerpene(c *gin.Context) {
-	var terpene models.Terpene
-	if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&terpene).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Terp not found!"})
-		return
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
+
+	if userIsLoggedIn {
+		var terpene models.Terpene
+		if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&terpene).Error; err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Terp not found!"})
+			return
+		}
+
+		// Validate input
+		var input terpeneInput
+		if err := c.ShouldBindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			return
+		}
+
+		terpene.Categories = append(terpene.Categories, input.Category...)
+		terpene.Smells = append(terpene.Smells, input.Smell...)
+		terpene.Tastes = append(terpene.Tastes, input.Taste...)
+		terpene.Effects = append(terpene.Effects, input.Effect...)
+		terpene.Strains = append(terpene.Strains, input.Strain...)
+
+		dbconnector.DB.Session(&gorm.Session{FullSaveAssociations: true}).Save(&terpene)
+
+		c.JSON(http.StatusOK, gin.H{"data": &terpene})
 	}
-
-	// Validate input
-	var input terpeneInput
-	if err := c.ShouldBindJSON(&input); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	terpene.Categories = append(terpene.Categories, input.Category...)
-	terpene.Smells = append(terpene.Smells, input.Smell...)
-	terpene.Tastes = append(terpene.Tastes, input.Taste...)
-	terpene.Effects = append(terpene.Effects, input.Effect...)
-	terpene.Strains = append(terpene.Strains, input.Strain...)
-
-	dbconnector.DB.Session(&gorm.Session{FullSaveAssociations: true}).Save(&terpene)
-
-	c.JSON(http.StatusOK, gin.H{"data": &terpene})
 }
 
 // DELETE /terpenes/:id
 // Delete a terpene
 func DeleteTerpene(c *gin.Context) {
-	var terpene models.Terpene
-	if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&terpene).Error; err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Terpene not found!"})
-		return
+	userIsLoggedIn := simpleauth.ValidateRequest(c)
+
+	if userIsLoggedIn {
+		var terpene models.Terpene
+		if err := dbconnector.DB.Where("id = ?", c.Param("id")).First(&terpene).Error; err != nil {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Terpene not found!"})
+			return
+		}
+
+		dbconnector.DB.Delete(&terpene)
+
+		c.JSON(http.StatusOK, gin.H{"data": true})
 	}
-
-	dbconnector.DB.Delete(&terpene)
-
-	c.JSON(http.StatusOK, gin.H{"data": true})
 }
